Move code around a little more.
I moved `QuadCoordinateConfig` up to the `geometry` module root, because it's
related to both quads and faces, and I literally forgot which file it was in
every single time I went looking for it.

I also move the quad buffer types out of `geometry` and into a `buffers` module.
